Rename the asset-specs.js file to assetSpecs.js
