+++ js/src/components/paid-ads/assetSpecs.js
